+++ src/Server/Process/Started.php
@@ -3,18 +3,18 @@ declare(strict_types = 1);
 
 namespace Innmind\Server\Control\Server\Process;
 
-use Innmind\Server\Control\{
-    Server\Process\Output\Chunk,
-    Server\Process\Output\Type,
-    Server\Signal,
+use Innmind\Server\Control\Server\{
+    Process\Output\Chunk,
+    Process\Output\Type,
+    Signal,
 };
 use Innmind\Filesystem\File\Content;
-use Innmind\TimeContinuum\{
+use Innmind\Time\{
     Clock,
-    PointInTime,
+    Point,
     Period,
+    Halt,
 };
-use Innmind\TimeWarp\Halt;
 use Innmind\IO\{
     IO,
     Streams\Stream,
@@ -44,7 +44,7 @@ final class Started
     private Halt $halt;
     private Period $grace;
     private bool $background;
-    private PointInTime $startedAt;
+    private Point $startedAt;
     /** @var resource */
     private $process;
     private Stream $output;
@@ -177,7 +177,7 @@ final class Started
             }
 
             /** @var Sequence<Either<ExitCode|'signaled'|'timed-out', SideEffect>> */
-            yield Sequence::of(Either::right(new SideEffect));
+            yield Sequence::of(Either::right(SideEffect::identity));
         })->flatMap(static fn($chunks) => $chunks);
     }
 
@@ -330,7 +330,7 @@ final class Started
     private function abort(): string
     {
         @\proc_terminate($this->process);
-        ($this->halt)($this->grace);
+        $_ = ($this->halt)($this->grace)->memoize();
 
         if ($this->status()['running']) {
             @\proc_terminate($this->process, Signal::kill->toInt());