+++ tests/ServerFactoryTest.php
@@ -6,10 +6,12 @@ namespace Tests\Innmind\Server\Control;
 use Innmind\Server\Control\{
     ServerFactory,
     Server,
-    Exception\UnsupportedOperatingSystem
+    Exception\UnsupportedOperatingSystem,
 };
-use Innmind\TimeContinuum\Clock;
-use Innmind\TimeWarp\Halt\Usleep;
+use Innmind\Time\{
+    Clock,
+    Halt,
+};
 use Innmind\IO\IO;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 use PHPUnit\Framework\Attributes\Group;
@@ -29,7 +31,7 @@ class ServerFactoryTest extends TestCase
         $this->assertInstanceOf(Server::class, ServerFactory::build(
             Clock::live(),
             IO::fromAmbientAuthority(),
-            Usleep::new(),
+            Halt::new(),
         ));
     }
 
@@ -48,7 +50,7 @@ class ServerFactoryTest extends TestCase
         ServerFactory::build(
             Clock::live(),
             IO::fromAmbientAuthority(),
-            Usleep::new(),
+            Halt::new(),
         );
     }
 }