+++ tests/Servers/UnixTest.php
@@ -4,13 +4,14 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Server\Control\Servers;
 
 use Innmind\Server\Control\{
-    Servers\Unix,
     Server,
     Server\Processes,
     Server\Volumes,
 };
-use Innmind\TimeContinuum\Clock;
-use Innmind\TimeWarp\Halt\Usleep;
+use Innmind\Time\{
+    Clock,
+    Halt,
+};
 use Innmind\IO\IO;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 use PHPUnit\Framework\Attributes\Group;
@@ -19,23 +20,12 @@ class UnixTest extends TestCase
 {
     #[Group('ci')]
     #[Group('local')]
-    public function testInterface()
-    {
-        $this->assertInstanceOf(Server::class, Unix::of(
-            Clock::live(),
-            IO::fromAmbientAuthority(),
-            Usleep::new(),
-        ));
-    }
-
-    #[Group('ci')]
-    #[Group('local')]
     public function testProcesses()
     {
-        $this->assertInstanceOf(Processes::class, Unix::of(
+        $this->assertInstanceOf(Processes::class, Server::new(
             Clock::live(),
             IO::fromAmbientAuthority(),
-            Usleep::new(),
+            Halt::new(),
         )->processes());
     }
 
@@ -43,10 +33,10 @@ class UnixTest extends TestCase
     #[Group('local')]
     public function testVolumes()
     {
-        $this->assertInstanceOf(Volumes::class, Unix::of(
+        $this->assertInstanceOf(Volumes::class, Server::new(
             Clock::live(),
             IO::fromAmbientAuthority(),
-            Usleep::new(),
+            Halt::new(),
         )->volumes());
     }
 }