+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 7.0.0 - 2026-01-31
+
+### Changed
+
+- Requires PHP `8.4`
+- `Innmind\Server\Control\Server` is now a final class, all previous implementations are now flagged as internal
+- `Innmind\Server\Control\Server\Volumes` is now a final class
+- `Innmind\Server\Control\Server\Processes` is now a final class
+- Requires `innmind/time:~1.0`
+
 ## 6.1.0 - 2025-08-06
 
 ### Added